Delete core/arch/arm/tee/tee_rpmb.c and core/include/tee/tee_rpmb.h
Move code into tee_rpmb_fs_common.c.

Tested-by: Pascal Brand <pascal.brand@linaro.org> (STM)
